Assignment 2 - Kubernetes App/Microservices_App/app/requirements.txt
grpcio
googleapis-common-protos
flask
pandas
pymongo